Add a version of n-queens that propagates constraints.
Set the board size (at the bottom of the file) to 20 and compare the speeds.  The first solution is always the longest. Subsequent solutions are faster depending on how much backtracking is required to find the next one.  Here are the times (in seconds on my computer) for the first 5 solutions using the three versions.

n_queens:                        76.332, 0.003,, 13.088, 6.502, 1.689
n_queens_no_logic_vars: 15.033, 0.001,    2.683, 1.497, 0.346
n_queens_cp:                    0.005, 0.004,    0.002, 0.005,  0.002